services/api-gateway/requirements.txt
fastapi==0.127.0
uvicorn==0.40.0
httpx==0.28.1
python-jose==3.5.0
python-dotenv==1.2.1
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.50.0
pytest==9.0.2
pytest-asyncio==1.3.0
